requirements.txt
# Core dependencies for ExamGenerator library
reportlab>=4.0.0
pillow>=10.0.0
qrcode>=7.0.0

# Optional: Only needed if using password-protected PDFs
# cryptography is included in reportlab for PDF encryption
# cffi and pycparser are transitive dependencies that will be installed automatically
